backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
requests>=2.31.0
nltk>=3.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
supabase>=2.7.0

